// types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
      Leitura: {
        Row: {
          end: string | null
          idLeitura: number
          idLivro: number
          pagFinal: number | null
          pagInicial: number | null
          start: string | null
        }
        Insert: {
          end?: string | null
          idLeitura?: number
          idLivro: number
          pagFinal?: number | null
          pagInicial?: number | null
          start?: string | null
        }
        Update: {
          end?: string | null
          idLeitura?: number
          idLivro?: number
          pagFinal?: number | null
          pagInicial?: number | null
          start?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_Leitura_Livro"
            columns: ["idLivro"]
            isOneToOne: false
            referencedRelation: "Livro"
            referencedColumns: ["idLivro"]
          },
        ]
      }
      Livro: {
        Row: {
          ano: number | null
          autor: string | null
          dtFinal: string | null
          dtInicial: string | null
          idLivro: number
          idUsuario: string
          img: string | null
          nome: string
          numPag: number
          numPagRead: number
          status: Database["public"]["Enums"]["status_livro"]
          text: string | null
        }
        Insert: {
          ano?: number | null
          autor?: string | null
          dtFinal?: string | null
          dtInicial?: string | null
          idLivro?: number
          idUsuario: string
          img?: string | null
          nome: string
          numPag: number
          numPagRead?: number
          status: Database["public"]["Enums"]["status_livro"]
          text?: string | null
        }
        Update: {
          ano?: number | null
          autor?: string | null
          dtFinal?: string | null
          dtInicial?: string | null
          idLivro?: number
          idUsuario?: string
          img?: string | null
          nome?: string
          numPag?: number
          numPagRead?: number
          status?: Database["public"]["Enums"]["status_livro"]
          text?: string | null
        }
        Relationships: []
      }
      profile: {
        Row: {
          id: string
          is_admin: boolean
        }
        Insert: {
          id?: string
          is_admin?: boolean
        }
        Update: {
          id?: string
          is_admin?: boolean
        }
        Relationships: []
      }
      push_subscriptions: {
        Row: {
          auth: string
          created_at: string
          endpoint: string
          id: string
          p256dh: string
          updated_at: string
          user_id: string | null
        }
        Insert: {
          auth: string
          created_at?: string
          endpoint: string
          id?: string
          p256dh: string
          updated_at?: string
          user_id?: string | null
        }
        Update: {
          auth?: string
          created_at?: string
          endpoint?: string
          id?: string
          p256dh?: string
          updated_at?: string
          user_id?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      status_livro: "Desejado" | "Lendo" | "Lido"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never